backend/requirements.txt
Flask==3.0.0
MongoEngine==0.27.0
pymongo==4.6.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.0
PyJWT==2.8.0
